Refactor and modularize image utilities; introduce gradient generator for tests and streamline color utilities.
